Added auto1_float as new model

+++ scripts/data.py
@@ -39,10 +39,10 @@ class PaintsTorchDataset(Dataset):
         H, W, C = buffer.shape
         y = np.random.randint(0, H - 512 + 1)
         x = np.random.randint(0, W - 512 + 1)
-        tensor = torch.from_numpy(buffer).permute(2, 0, 1).to(torch.bfloat16)
+        tensor = torch.from_numpy(buffer).permute(2, 0, 1).float()
         tensor = tensor[:, y:y + 512, x:x + 512]
         return Triplet(
-            pre_illustration(tensor[:3].to(torch.bfloat16)),
-            pre_lineart(tensor[3 + np.random.randint(0, 3), None].to(torch.bfloat16)),
-            pre_illustration(tensor[6:].to(torch.bfloat16)),
+            pre_illustration(tensor[:3].float()),
+            pre_lineart(tensor[3 + np.random.randint(0, 3), None].float()),
+            pre_illustration(tensor[6:].float()),
         )
